test: add update StatuteSelector unit tests
